Keep state in the original object instead of copying in/out.
Previously each Object/List/FieldState could copy values out of the
input object, and then re-create the input object on each call to
`.value`.

Now we keep a stable identity of the currently-editing object, and
just read/write the values directly out of that object.

I'm a little surprised this worked, but so far so good.